# server/src/configs/openapi.yaml
openapi: 3.0.3
info:
  title: CloudDrive16 API
  description: |-
    This project was used to build CloudDrive16 web (A file hosting service and synchronization service operated by CloudTeam16)
  termsOfService: http://swagger.io/terms/
  contact:
    email: luclegen.info@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:5000/api
...
paths:
  /users:
    post:
      tags:
        - users
      summary: Create user
      description: Create a new user
      operationId: createUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: "Registered successfully."
        '400':
          description: "1. Invalid email;\n2. Invalid language;\n3. Invalid role.|\nPlease choose a stronger password. Try a mix of letters, numbers, and symbols (use 8 or more characters)."
        '403':
          description: "Invalid birthday."
        '404':
          description: "User not found.|Profile not found."
        '422':
          description: "Email is taken!"
      security:
        - petstore_auth:
            - write:pets
            - read:pets
    get:
      tags:
        - users
      summary: Read user
      description: Read an user 
      operationId: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: boolean
                default: true
        '203':
          description: failure operation
          content:
            application/json:
              schema:
                type: boolean
                default: false
        '400':
          description: "Invalid email."
        '401':
          description: "Unauthorized."
      security:
        - petstore_auth:
            - write:pets
            - read:pets
    put:
      tags:
        - users
      summary: Update user
      description: Update user
      operationId: updateUser
      requestBody:
        description: Update user in CloudDrive16
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdatedUser'
        required: true
      responses:
        '200':
          description: "Updated successfully."
        '403':
          description: "Cannot execute \"update\" on \"Profile\"."
        '404':
          description: "User not found.|Profile not found."
      security:
        - petstore_auth:
            - write:pets
            - read:pets
    patch:
      tags:
        - users
      summary: Change language
      description: Change language of user
      operationId: changeLang
      requestBody:
        description: Change lang
        content:
          text/plain:
            schema:
              type: string
              example: en|vi
      responses:
        '200':
          description: "Change language successfully."
        '400':
          description: "1. Invalid language."
        '403':
          description: "Cannot execute \"read\" on \"User\".|Cannot execute \"changeLang\" on \"User\""
        '404':
          description: "User not found."
...
components:
  schemas:
...
    User:
      type: object
      properties:
        first_name:
          type: string
          example: Luc
        last_name:
          type: string
          example: Tan Huynh
        email:
          type: string
          example: luclegen.dev@gmail.com
        password:
          type: string
          example: Luc123!@#
        day:
          type: integer
          example: 29
        month:
          type: integer
          example: 9
        year:
          type: integer
          example: 1997
        sex:
          type: string
          enum:
            - Male
            - Female
            - Other
          default: Male
        lang:
          type: string
          enum:
            - en
            - vi
          example: en|vi
      xml:
        name: user
    UpdatedUser:
      type: object
      properties:
        avatar:
          type: string
          format: binary
        first_name:
          type: string
          example: Liên
        middle_name:
          type: string
          example: Thị Cẩm
        last_name:
          type: string
          example: Huỳnh
        lang:
          type: string
          enum:
            - en
            - vi
          default: en
        full_name:
          type: string
          example: Huỳnh Thị Cẩm Liên
        day:
          type: integer
          example: 16
        month:
          type: integer
          example: 2
        year:
          type: integer
          example: 1990
        sex:
          type: string
          enum:
            - Male
            - Female
            - Other
          default: Female
      xml:
        name: user
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header